Update Metabase Functions.R
@MargheritaPhilipp @cabrinenyona 
Sorry I could not make the meeting this morning. I found the bug - I was accidentally not merging in all the fuzzyjoin matches. I have fixed it now. I've also added in a warning message so the user knows that we are merging in fuzzyjoin matches!

Cabrine - thank you again for pointing out the bug!